Authentication is robust now (at least on my machine)
I just connected to 2b2t 10 times in a row without any issues.
Not sure about the mess of a solution and if it should be proposed upstream.
I have the feeling that the problem I fixed here does not exist for others.
So maybe I did something fundamentally wrong here.

But whatever it is working stable on my machine now :shrug: